pyproject.toml
[tool.poetry]
name = "mydjangoproject"
version = "0.1.0"
description = ""
authors = ["XyzzyZ4p <xyzzyzap@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
python-dotenv = "^0.21.1"
psycopg2 = "^2.9.5"
uvicorn = "^0.20.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
django-extensions = "^3.2.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
